requirements.txt
google-cloud-texttospeech>=2.21.0
python-dotenv>=1.0.0
